Update stickymate.js
fix `animation.elements.forEach & classes.elements.forEach is not a function.`

If you have bundled stickymate in your code and run it on a page for example, where NO sticky element is found, it will break the whole javascript (since the script always assume at least one data-sticky. This simple check fixes this regression introduced in 1.3.0.