the link to npm is removed
Its become too much complicated to keep the package on npm in sync with GitHub version. Since I moved to new Macs, it was required to reinstall brew and node.js again, remember all the commands and so on. For the last 3 years I specializes on Type Design and have not so much time to do all this things again, and npm definitely need to provide an easier way to update packages. So I'm sorry but for now stickymate will be available only here on GitHub.